+++ app/login/components/color-palettes/color-palette-types.tsx
@@ -246,7 +246,6 @@ const CategoricalColorPaletteCreator = (props: ColorPaletteCreatorProps) => {
           ))}
         </Box>
       )}
-
       <Button
         variant="text"
         fullWidth